requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
black==25.9.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.11.0
cryptography==46.0.2
daphne==4.2.1
distro==1.9.0
Django==5.2.7
django-tinymce==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.11
flake8==7.3.0
googlemaps==4.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mccabe==0.7.0
msgpack==1.1.2
mypy==1.18.2
mypy_extensions==1.1.0
openai==2.6.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==13.0.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.0.1
